bootstrapping is now a config operation
- feature request to make late bootstrapping easier
