+++ configerus/config.py
@@ -1,10 +1,14 @@
 import logging
+from importlib import metadata
 
 from .plugin import Factory, Type, PluginInstances
 from .shared import tree_merge
 from .loaded import Loaded
 
 logger = logging.getLogger('configerus:config')
+
+CONFIGERUS_BOOTSTRAP_ENTRYPOINT="configerus.bootstrap"
+""" SetupTools entrypoint target for bootstrapping """
 
 CONFIG_PATH_LABEL = 'paths'
 """ If you load this label, it is meant to be return a keyed path """
@@ -65,6 +69,33 @@ class Config:
         copy = Config()
         copy.plugins = self.plugins.copy(copy)
         return copy
+
+    def bootstrap(self, bootstrap_id:str):
+        """
+        Make a plugin object instance of a type and key
+
+        A python module out there will need to have declared an entry_point for
+        'mirantis.testing.toolbox.{type}' with entry_point key {name}.
+        The entrypoint must be a factory method of signature:
+        ```
+            def XXXX(conf: configerus.config.Config, instance_id:str) -> {plugin}:
+        ```
+
+        An alternative would be if the entrypoint refers to a sub-package here. We
+        directly
+
+        The factory should return the plugin, which is likely going to be some kind
+        of an object which has value to the caller of the function. This function
+        does not specify what the return needs to be.
+        """
+        logger.info("Running bootstrap entrypoint: %s", bootstrap_id)
+        eps = metadata.entry_points()[CONFIGERUS_BOOTSTRAP_ENTRYPOINT]
+        for ep in eps:
+            if ep.name == bootstrap_id:
+                plugin = ep.load()
+                return plugin(self)
+        else:
+            raise KeyError("Bootstrap not found {}.{}".format(bootstrap_id, CONFIGERUS_BOOTSTRAP_ENTRYPOINT))
 
     """ Pass through accessors
